refactor: pages 폴더명 케밥케이스로 변경 및 PAGE_URL 상수 작성
