+++ src/constants/pageUrl.ts
@@ -1,0 +1,4 @@
+export const PAGE_URL = {
+  ADD_CARD: "/add-card",
+  ADD_CARD_SUCCES: "/add-card/success",
+};
