+++ src/pages/add-card/paymentInput/components/cardInputForm/cardInput/CardNumberInput.tsx
@@ -4,7 +4,7 @@ import Input from "../../../../../../components/common/inputForm/input/Input";
 import { CARD_INFO } from "../../constants/CardInfo";
 import { validateNumberString } from "./validator/validateCardInput";
 import { getFirstErrorMessage } from "./validator/getFirstErrorMessage";
-import { CardInfo } from "../../../../paymentInput/PaymentInputPage";
+import { CardInfo } from "../../../PaymentInputPage";
 
 function CardNumberInput({
   setCardInfo,
