+++ src/pages/add-card/paymentInput/components/cardInputForm/cardInput/validator/validateCardInput.ts
