+++ src/pages/add-card/paymentInput/components/constants/CardInfo.ts
