+++ src/stories/CardExpirationDateInput.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import CardExpirationDateInput from "../pages/addCard/components/paymentInputPage/cardInputForm/cardInput/CardExpirationDateInput";
+import CardExpirationDateInput from "../pages/add-card/components/paymentInputPage/cardInputForm/cardInput/CardExpirationDateInput";
 import { expect, userEvent, within } from "@storybook/test";
 import styles from "../components/common/inputForm/input/Input.module.css";
