+++ src/stories/CardNumberInput.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import { useState } from "react";
-import CardNumberInput from "../pages/addCard/components/paymentInputPage/cardInputForm/cardInput/CardNumberInput";
+import CardNumberInput from "../pages/add-card/components/paymentInputPage/cardInputForm/cardInput/CardNumberInput";
 import { userEvent, expect, within } from "@storybook/test";
 import styles from "../components/common/inputForm/input/Input.module.css";
